Fix: interfaces
